refactor: update path to the schema
