+++ src/pages/forgot-password/model/validation-schema.ts
@@ -1,6 +1,6 @@
 import { z } from 'zod';
 
-import { emailSchema } from '@shared/schemas';
+import { emailSchema } from '@shared/lib/schemas';
 
 export const forgotPasswordSchema = z.object({
   email: emailSchema,
