+++ src/pages/reset-password/model/validation-schema.ts
@@ -1,6 +1,6 @@
 import { z } from 'zod';
 
-import { confirmPasswordSchema } from '@shared/schemas';
+import { confirmPasswordSchema } from '@shared/lib/schemas';
 import { passwordSchema } from '@shared/ui/password-field/model/validation-schema';
 
 export const resetPasswordSchema = z
